Refactoring + new test
- moving implementations from DefaultAcLoadFlowObserver.java to a new class named NetworkEquationSystemAcLoadFlowObserver.java
- Limites: Vérifier que le switch se fait, ie qu'on passe bien en PQ si le SVC atteint QMin ou QMax.
